Rename datahandler module to history module
Also move previous history.py into this module - so everything is
bundled
