+++ freqtrade/data/history/__init__.py
@@ -1,0 +1,14 @@
+"""
+Handle historic data (ohlcv).
+
+Includes:
+* load data for a pair (or a list of pairs) from disk
+* download data from exchange and store to disk
+"""
+
+from .history_utils import (convert_trades_to_ohlcv,  # noqa: F401
+                            get_timerange, load_data, load_pair_history,
+                            refresh_backtest_ohlcv_data,
+                            refresh_backtest_trades_data, refresh_data,
+                            validate_backtest_data)
+from .idatahandler import get_datahandler, get_datahandlerclass  # noqa: F401
