+++ freqtrade/data/history/jsondatahandler.py
